+++ src/app/shared/toastr/toastr.component.ts
@@ -79,7 +79,6 @@ export class ToastrComponent implements OnInit, OnDestroy {
       case 'success':
         iconHtml = `
           <div class="flex items-center">
-            <span class="text-xl mr-1">✅</span>
             <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
               <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 13l4 4L19 7" />
             </svg>
@@ -89,7 +88,6 @@ export class ToastrComponent implements OnInit, OnDestroy {
       case 'error':
         iconHtml = `
           <div class="flex items-center">
-            <span class="text-xl mr-1">❌</span>
             <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
               <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12" />
             </svg>
@@ -99,7 +97,6 @@ export class ToastrComponent implements OnInit, OnDestroy {
       case 'warning':
         iconHtml = `
           <div class="flex items-center">
-            <span class="text-xl mr-1">⚠️</span>
             <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
               <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z" />
             </svg>
@@ -109,7 +106,6 @@ export class ToastrComponent implements OnInit, OnDestroy {
       case 'info':
         iconHtml = `
           <div class="flex items-center">
-            <span class="text-xl mr-1">ℹ️</span>
             <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
               <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M13 16h-1v-4h-1m1-4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
             </svg>